refactor short circuit api
