+++ pypowsybl/loadflow/impl/loadflow.py
@@ -1,4 +1,3 @@
-#
 # Copyright (c) 2020-2022, RTE (http://www.rte-france.com)
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
