+++ pypowsybl/network/impl/node_breaker_topology.py
@@ -1,3 +1,9 @@
+# Copyright (c) 2023, RTE (http://www.rte-france.com)
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+# SPDX-License-Identifier: MPL-2.0
+#
 from pandas import DataFrame
 import networkx as _nx
 import pypowsybl._pypowsybl as _pp
